spec+hints: minor cleanups — stub-status, piece_count width
- `randomizer-logic/spec.md` (trick logic delta): added `> **Stub
  status**` blockquote to the `Per-trick ROM-version verification
  status` requirement so readers can distinguish landed-and-tested
  from aspirational-and-tracked. Pattern matches the other Phase B
  delta requirements that use the same blockquote shape.
- `rando_hints.c`: widened `piece_count` from uint8 to uint16. The
  `settings.pieces_placed` axis is uint16 (range up to 65535) and a
  uint8 counter would wrap silently past 255 — practical Triforce
  Hunt range is ~50 today but the type pin guards against future
  axis widening.

Both LOW-priority cluster-audit findings; bundled because they're
small spec/header cleanups with no behavioral change.

+++ src/rando/rando_hints.c
@@ -215,7 +215,11 @@ bool Rando_GenerateHints(const RandoSettings *settings,
     // be crossed in a future slice). Audit-of-audit LOW-2 of phase-b.
     uint16 seen_regions[16] = {0};
     uint8 seen_count = 0;
-    uint8 piece_count = 0;
+    uint16 piece_count = 0;  // uint16 — `settings.pieces_placed` is uint16
+                              // (range up to 65535); a uint8 here would wrap
+                              // silently past 255. Practical TH max is ~50
+                              // but the type pin guards against future axis
+                              // widening.
     for (uint16 i = 0; i < entry_count; i++) {
       if (placements->entries[i].item_id != ITEM_TriforcePiece) continue;
       piece_count++;
